requirements.txt
-i https://pypi.org/simple
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
defusedxml==0.5.0; python_version >= '3.0'
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-redis-cache==1.8.1
django==2.1.4
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
kombu==4.2.2
lxml==4.2.5
markdown==3.0.1
oauthlib==2.1.0
psycopg2-binary==2.7.6.1
pyjwt==1.7.1
python3-openid==3.1.0; python_version >= '3.0'
pytz==2018.7
redis==2.10.6
requests-oauthlib==1.0.0
requests==2.21.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
vine==1.1.4
